Polish the CPU profiler status view
