+++ packages/devtools_app/test/shared/common_widgets_test.dart
@@ -10,9 +10,6 @@ import 'package:flutter_test/flutter_test.dart';
 // TODO(kenz): add tests for other widgets in common_widgets.dart
 
 void main() {
-  const instructionsKey = Key('instructions');
-  const recordingStatusKey = Key('recordingStatus');
-  const processingStatusKey = Key('processingStatus');
   const windowSize = Size(1000.0, 1000.0);
 
   setUp(() {
@@ -23,93 +20,6 @@ void main() {
   });
 
   group('Common widgets', () {
-    testWidgetsWithWindowSize('recordingInfo builds info for pause', windowSize,
-        (WidgetTester tester) async {
-      await tester.pumpWidget(
-        wrap(
-          const RecordingInfo(
-            instructionsKey: instructionsKey,
-            recordingStatusKey: recordingStatusKey,
-            processingStatusKey: processingStatusKey,
-            recording: false,
-            recordedObject: 'fake object',
-            processing: false,
-            isPause: true,
-          ),
-        ),
-      );
-
-      expect(find.byKey(instructionsKey), findsOneWidget);
-      expect(find.byKey(recordingStatusKey), findsNothing);
-      expect(find.byKey(processingStatusKey), findsNothing);
-      expect(find.text('Click the pause button '), findsOneWidget);
-      expect(find.text('Click the stop button '), findsNothing);
-    });
-
-    testWidgetsWithWindowSize('recordingInfo builds info for stop', windowSize,
-        (WidgetTester tester) async {
-      await tester.pumpWidget(
-        wrap(
-          const RecordingInfo(
-            instructionsKey: instructionsKey,
-            recordingStatusKey: recordingStatusKey,
-            processingStatusKey: processingStatusKey,
-            recording: false,
-            recordedObject: 'fake object',
-            processing: false,
-          ),
-        ),
-      );
-
-      expect(find.byKey(instructionsKey), findsOneWidget);
-      expect(find.byKey(recordingStatusKey), findsNothing);
-      expect(find.byKey(processingStatusKey), findsNothing);
-      expect(find.text('Click the stop button '), findsOneWidget);
-      expect(find.text('Click the pause button '), findsNothing);
-    });
-
-    testWidgetsWithWindowSize(
-        'recordingInfo builds recording status', windowSize,
-        (WidgetTester tester) async {
-      await tester.pumpWidget(
-        wrap(
-          const RecordingInfo(
-            instructionsKey: instructionsKey,
-            recordingStatusKey: recordingStatusKey,
-            processingStatusKey: processingStatusKey,
-            recording: true,
-            recordedObject: 'fake object',
-            processing: false,
-          ),
-        ),
-      );
-
-      expect(find.byKey(instructionsKey), findsNothing);
-      expect(find.byKey(recordingStatusKey), findsOneWidget);
-      expect(find.byKey(processingStatusKey), findsNothing);
-    });
-
-    testWidgetsWithWindowSize(
-        'recordingInfo builds processing status', windowSize,
-        (WidgetTester tester) async {
-      await tester.pumpWidget(
-        wrap(
-          const RecordingInfo(
-            instructionsKey: instructionsKey,
-            recordingStatusKey: recordingStatusKey,
-            processingStatusKey: processingStatusKey,
-            recording: false,
-            recordedObject: 'fake object',
-            processing: true,
-          ),
-        ),
-      );
-
-      expect(find.byKey(instructionsKey), findsNothing);
-      expect(find.byKey(recordingStatusKey), findsNothing);
-      expect(find.byKey(processingStatusKey), findsOneWidget);
-    });
-
     testWidgetsWithWindowSize(
         'processingInfo builds for progressValue', windowSize,
         (WidgetTester tester) async {
